chore: update Custody ui
